4-FastAPI con JWT/requirements.txt
fastapi[standard]
passlib[bcrypt]
PyJWT
